% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnect.R
\name{dbConnect}
\alias{dbConnect}
\title{Connect to the SQL database}
\usage{
dbConnect(db.info)
}
\arguments{
\item{db.info}{A list with fields: user, password, dbname, and host}
}
\description{
Connect to the SQL database
}
