% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Get gene information}
\usage{
getGeneInfo(gene, con = NULL, db.info)
}
\arguments{
\item{gene}{HGNC gene symbol}
}
\value{
data.frame of gene info
}
\description{
Get gene information
}
\examples{
getGeneInfo('PCSK9')
}
