% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotation.R
\name{getVariantAnnotation}
\alias{getVariantAnnotation}
\title{Fetches data from HGMD, OMIM, CLINVAR, and the variant summary table-materialized_view (variant QC stats, 
allele frequencies. and functional predictions.)}
\usage{
getVariantAnnotation(chrom, pos, ref, alt, con = NULL, db.info)
}
\arguments{
\item{chrom}{Chromosome, as string, without the "chr" prefix}

\item{pos}{Position of the SNP. If structural variant, it's the start.}

\item{ref}{Reference allele}

\item{alt}{Alternative allele}
}
\description{
Fetches data from HGMD, OMIM, CLINVAR, and the variant summary table-materialized_view (variant QC stats, 
allele frequencies. and functional predictions.)
}
