% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantInfo.R
\name{getVariantInfo}
\alias{getVariantInfo}
\title{query DB tables for variant informaiton}
\usage{
getVariantInfo(
  snps,
  varids,
  source = names(.src.to.dbtable),
  con = NULL,
  db.info
)
}
\arguments{
\item{snps}{Vector of SNPs in the format chr_pos_ref_alt}

\item{varids}{Vector of variant IDs}

\item{source}{Name of the dataset to pull from}
}
\description{
query DB tables for variant informaiton
}
\examples{
getVariantInfo(varids=c(1221787, 1221983, 1222865), source='medvardb')
}
