% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantsByGene.R
\name{getVariantsByGene}
\alias{getVariantsByGene}
\title{Get variants in a gene}
\usage{
getVariantsByGene(
  gene,
  vep_var_impact = NULL,
  canonical = TRUE,
  con = NULL,
  db.info = .fah.dbinfo
)
}
\arguments{
\item{gene}{HGNC gene symbol}

\item{vep_var_impact}{One of {HIGH, LOW, MODERATE, MODIFIER, NA}}

\item{canonical}{If true, limit to canonical transcipts}
}
\description{
Get variants in a gene
}
\examples{
get.variants.by.gene(gene='PCSK9', vep_var_impact='HIGH', canonical=TRUE)
}
