% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDbConnect.R
\name{myDbConnect}
\alias{myDbConnect}
\title{Connect to the SQL database}
\usage{
myDbConnect(db.info)
}
\arguments{
\item{db.info}{A list with fields: user, password, dbname, and host}
}
\description{
Connect to the SQL database
}
